refactor:: yunNote 반영 리팩토링
