+++ module-admin/src/main/java/com/kernel360/product/service/ProductService.java
@@ -8,7 +8,7 @@ import com.kernel360.product.dto.ProductDto;
 import com.kernel360.product.dto.ProductUpdateRequest;
 import com.kernel360.product.dto.RecommendProductsDto;
 import com.kernel360.product.entity.Product;
-import com.kernel360.product.entity.SafetyStatus;
+import com.kernel360.product.enumset.SafetyStatus;
 import com.kernel360.product.repository.ProductRepositoryJpa;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
