+++ module-admin/src/main/resources/application.yml
@@ -5,6 +5,10 @@ spring:
     multipart:
       max-file-size: 50MB
       max-request-size: 500MB
+  jpa:
+    hibernate:
+      naming:
+        physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
 server:
   port: 8082
