+++ module-api/src/main/java/com/kernel360/main/controller/MainController.java
@@ -3,6 +3,7 @@ package com.kernel360.main.controller;
 import com.kernel360.main.code.BannerBusinessCode;
 import com.kernel360.main.dto.BannerDto;
 import com.kernel360.main.dto.RecommendProductsDto;
+import com.kernel360.main.enumset.Sort;
 import com.kernel360.main.service.MainService;
 import com.kernel360.product.code.ProductsBusinessCode;
 import com.kernel360.product.dto.ProductResponse;
