+++ module-batch/src/main/java/com/kernel360/modulebatch/product/job/infra/ConcernedProductToProductListItemProcessor.java
@@ -6,7 +6,7 @@ import com.kernel360.ecolife.entity.ConcernedProduct;
 import com.kernel360.ecolife.repository.ConcernedProductRepository;
 import com.kernel360.modulebatch.product.dto.ProductDto;
 import com.kernel360.product.entity.Product;
-import com.kernel360.product.entity.SafetyStatus;
+import com.kernel360.product.enumset.SafetyStatus;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
