+++ module-domain/src/main/java/com/kernel360/washinfo/entity/WashInfo.java
@@ -24,20 +24,15 @@ public class WashInfo extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "wash_info_id_gen")
     @SequenceGenerator(name = "wash_info_id_gen", sequenceName = "wash_info_wash_no_seq", allocationSize = 50)
-    @Column(name = "wash_no", nullable = false)
+    @Column(nullable = false)
     private Integer washNo;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_no", nullable = false)
     private Member member;
 
-    @Column(name = "wash_count")
     private Integer washCount;
-
-    @Column(name = "monthly_expense")
     private Integer monthlyExpense;
-
-    @Column(name = "interest")
     private Integer interest;
 
     private WashInfo(Integer washCount, Integer monthlyExpense, Integer interest) {
